pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "document-portal"
version = "0.1.0"
description = "A document processing and management portal"
requires-python = ">=3.9"
authors = [
    {name = "Rohit More", email = "rohitmore@example.com"},
]

dependencies = [
    "faiss-cpu>=1.11.0.post1",
    "ipykernel>=6.29.5",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.1.8",
    "langchain-groq>=0.3.6",
    "pandas>=2.3.1",
    "pydantic>=2.11.7",
    "pymupdf>=1.26.3",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.2",
    "setuptools>=80.9.0",
    "streamlit>=1.48.0",
    "structlog>=25.4.0",
    "uuid>=1.30",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "ruff>=0.1.0",
]

[tool.ruff]
target-version = "py39"
line-length = 88
select = ["E", "W", "F", "I"]
ignore = ["E501"]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = ["--strict-markers"]
